+++ data/cli_options.js
@@ -1,10 +1,14 @@
-
 'use strict';
 
 module.exports = {
-  env: {
-    alias: 'e',
-    type: 'number',
-    desc: 'Specify current environment. '
-  }
+    env: {
+        alias: 'e',
+        type: 'number',
+        desc: 'Specify current environment. '
+    },
+    local: {
+        alias: 'l',
+        type: 'string',
+        desc: 'Specify current local value. '
+    }
 };