+++ project_config/webpack/static_server_url.js
@@ -1,3 +1,6 @@
+
+const _ = require('lodash');
+
 /**
  * handle staticServerUrl
  *